Delete the curtin-vmtest-sync-images job
The vmtests already sync the needed images before running, so this job
is not really needed.

As we plan to run the vmtests across a number of arches a single job
wouldn't have worked well, as Jenkins runs it on a single node, while
we'd like to update the images on all the nodes.