admin-jobs-update: improve the directory traversal
This loop:

  for dir in `find * -type d`

is flawed, see the first example in [1]. Better to use a simple glob, as
suggested there. As implemented by this commit the loop will only go
over first-level directories; we make JJB recurse in subdirectories by
calling jenkins-jobs with `--recursive`.

jenkins-jobs --recursive is not fully equivalent to passing every single
directory to jenkins-jobs, as with --recursive all the YAMLs are used
together to generate the jobs. This per-first-level-directory (i.e.
per-project) subdivision is in my opinion a good balance between
isolation between projects and the ability to have per-project common
JJB definitions. In the specific case of server-jenkins-jobs there is no
difference, as we don't have subdirectories.

[1] https://mywiki.wooledge.org/BashPitfalls