admin-jobs-update: support passing --delete-all to JJB
Add a parameter to optionally make JJB delete the jobs it used to manage
but doesn't manage anymore (obsolete jobs, see [1]). Useful when Ubuntu
releases reach EOL and several jobs are to be removed at the same time.

Also: make JJB run on all the jobs directories even if it fails running
in one directory. In this case report failure on exit.

[1] https://docs.openstack.org/infra/jenkins-job-builder/execution.html#deleting-jobs-views