git-ubuntu: directly notify the maintainers on failures
This time as a Mailer plugin pipeline step [1]. Note that the list
of recipients has to be comma-separated, even if the documentation
says space should also work as a separator (it works when using a
"freestyle" project publisher, but not in a pipeline).

To avoid duplication the mail recipients could be set in a variable by
converting the job definitions to job-templates (escaping all the curly
braces) and configuring a JJB project. Probably not worth it here.

[1] https://www.jenkins.io/doc/pipeline/steps/mailer/#stepclass-mailer-e-mail-notification
[2] https://plugins.jenkins.io/mailer/